+++ include/blade/channelizer/test.hh
@@ -14,8 +14,8 @@ public:
 
     Result process();
 
-    std::span<const std::complex<int8_t>> getInputData();
-    std::span<const std::complex<int8_t>> getOutputData();
+    std::span<const std::complex<float>> getInputData();
+    std::span<const std::complex<float>> getOutputData();
 };
 
 } // namespace Blade::Channelizer::Test
